+++ lending/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [typescript v4.0.0] lending
 ### Releases
 - [NPM v4.0.0] https://www.npmjs.com/package/@codat/lending/v/4.0.0 - lending
+
+## 2023-12-04 01:30:09
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.1] lending
+### Releases
+- [NPM v4.0.1] https://www.npmjs.com/package/@codat/lending/v/4.0.1 - lending